+++ lib/src/core/selection_controller_base.dart
@@ -3,6 +3,6 @@ part of '../../selection_group.dart';
 abstract interface class SelectionControllerBase<T> implements Listenable {
   void _register(T value, FocusNode node);
   void _unregister(T value);
-  void select(T value);
+  void select(T value, {bool fromPress = false});
   bool isSelected(T value);
 }